Fix Spree::Shipment#item_cost for split shipments
This commit is going to fix the bug reported here #2919.

Now when shipment is split, it creates two shipment but associated
line_item has quantity 2 which is resulting in incorrect item_cost
because it is taken from line_item not from shipment.
Shipment manifest know the right quantity for every shipment, use it to
calculate the item_cost as sum of

`(single_line item + weighted adjustment per line item) * quantity`